pyproject.toml
[tool.poetry]
name = "applying-ai-project"
version = "0.1.0"
description = ""
authors = ["Hannu Karstu <hannukarstu1@kamk.fi>"]
readme = "README.md"
packages = [{include = "applying_ai_project"}]

[tool.poetry.dependencies]
python = "3.8.10"
flask = "^2.2.2"
redis = "^4.4.0"
imageio = "^2.23.0"
opencv-python = "4.6.0.66"
matplotlib = "^3.6.2"
tensorflow-hub = "^0.12.0"
mediapipe = "^0.9.0.1"
pandas = "^1.5.2"
tensorflow-docs = {git = "https://github.com/tensorflow/docs"}
ipython = "^8.7.0"
tensorflow-cpu = "^2.11.0"
flask-pymongo = "2.2.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.9"
pytest = "^7.2.0"
autopep8 = "^2.0.1"
pytest-env = "^0.8.1"
ipykernel = "^6.19.4"
mongomock = "3.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
